+++ sunkit_image/utils/tests/test_utils.py
@@ -5,16 +5,11 @@ import astropy.units as u
 from astropy.tests.helper import assert_quantity_allclose
 
 from sunkit_image.utils import (
-    bandpass_filter,
     bin_edge_summary,
-    curvature_radius,
     equally_spaced_bins,
-    erase_loop_in_image,
     find_pixel_radii,
     get_radial_intensity_summary,
-    initial_direction_finding,
-    loop_add,
-    smooth,
+
 )
 
 
@@ -126,219 +121,3 @@ def test_find_pixel_radii(smap):
 def test_get_radial_intensity_summary():
     # TODO: Write some tests.
     pass
-
-
-@pytest.fixture
-def test_map():
-    map_test = [
-        [1.0, 1.0, 1.0, 1.0],
-        [1.0, 5.0, 5.0, 1.0],
-        [1.0, 5.0, 5.0, 1.0],
-        [1.0, 1.0, 1.0, 1.0],
-    ]
-
-    return np.array(map_test)
-
-
-@pytest.fixture
-def image():
-    return np.ones((4, 4), dtype=np.float32)
-
-
-def test_bandpass_filter(image, test_map):
-
-    expect = np.zeros((4, 4))
-    result = bandpass_filter(image)
-
-    assert np.allclose(expect, result)
-
-    expect = np.array(
-        [
-            [0.0, 0.0, 0.0, 0.0],
-            [0.0, 2.22222222, 2.22222222, 0.0],
-            [0.0, 2.22222222, 2.22222222, 0.0],
-            [0.0, 0.0, 0.0, 0.0],
-        ]
-    )
-
-    result = bandpass_filter(test_map)
-
-    assert np.allclose(expect, result)
-
-    with pytest.raises(ValueError) as record:
-        _ = bandpass_filter(image, 5, 1)
-
-    assert str(record.value) == "nsm1 should be less than nsm2"
-
-
-def test_smooth(image, test_map):
-
-    filtered = smooth(image, 1)
-    assert np.allclose(filtered, image)
-
-    filtered = smooth(image, 4)
-    assert np.allclose(filtered, image)
-
-    filtered = smooth(test_map, 1)
-    assert np.allclose(filtered, test_map)
-
-    filtered = smooth(test_map, 3)
-    expect = np.array(
-        [
-            [1.0, 1.0, 1.0, 1.0],
-            [1.0, 2.77777777, 2.77777777, 1.0],
-            [1.0, 2.77777777, 2.77777777, 1.0],
-            [1.0, 1.0, 1.0, 1.0],
-        ]
-    )
-
-    assert np.allclose(filtered, expect)
-
-
-def test_erase_loop_in_image(image, test_map):
-
-    # The starting point of a dummy loop
-    istart = 0
-    jstart = 1
-    width = 1
-
-    # The coordinates of the dummy loop
-    xloop = [1, 2, 3]
-    yloop = [1, 1, 1]
-
-    result = erase_loop_in_image(image, istart, jstart, width, xloop, yloop)
-
-    expect = np.array(
-        [[0.0, 0.0, 0.0, 1.0], [0.0, 0.0, 0.0, 1.0], [0.0, 0.0, 0.0, 1.0], [0.0, 0.0, 0.0, 1.0]]
-    )
-
-    assert np.allclose(expect, result)
-
-    result = erase_loop_in_image(test_map, istart, jstart, width, xloop, yloop)
-
-    expect = np.array(
-        [[0.0, 0.0, 0.0, 1.0], [0.0, 0.0, 0.0, 1.0], [0.0, 0.0, 0.0, 1.0], [0.0, 0.0, 0.0, 1.0]]
-    )
-
-    assert np.allclose(expect, result)
-
-
-@pytest.fixture
-def test_image():
-
-    # An image containing a loop in a straight line
-    ima = np.zeros((3, 3), dtype=np.float32)
-    ima[0, 1] = 5
-    ima[1, 1] = 3
-    ima[2, 1] = 0
-    return ima
-
-
-def test_initial_direction_finding(test_image):
-
-    # The starting point of the loop i.e. the maximumflux position
-    xstart = 0
-    ystart = 1
-    nlen = 30
-
-    # The angle returned is with respect to the ``x`` axis.
-    al = initial_direction_finding(test_image, xstart, ystart, nlen)
-
-    # The angle returned is zero because the image has loop in the ``y`` direction but the function
-    # assumes the image is transposed so it takes the straight line in the ``x`` direction.
-    assert np.allclose(al, 0.0)
-
-
-def test_curvature_radius(test_image):
-
-    xl = np.zeros((3), dtype=np.float32)
-    yl = np.zeros((3), dtype=np.float32)
-    zl = np.zeros((3), dtype=np.float32)
-    al = np.zeros((3), dtype=np.float32)
-    ir = np.zeros((3), dtype=np.float32)
-
-    xl[0] = 0
-    yl[0] = 1
-    zl[0] = 5
-    al[0] = 0.0
-
-    # Using the similar settings in as in the IDL tutorial.
-    # This is forward tracing where the first point is after the starting point is being traced.
-    xl, yl, zl, al = curvature_radius(test_image, 30, xl, yl, zl, al, ir, 0, 30, 0)
-
-    assert np.allclose(np.ceil(xl[1]), 1)
-    assert np.allclose(np.ceil(yl[1]), 1)
-    assert np.allclose(zl[1], 3)
-
-    # This is forward tracing where the second point is after the starting point is being traced.
-    xl, yl, zl, al = curvature_radius(test_image, 30, xl, yl, zl, al, ir, 1, 30, 0)
-
-    assert np.allclose(np.ceil(xl[2]), 2)
-    assert np.allclose(np.ceil(yl[2]), 1)
-    assert np.allclose(zl[2], 0)
-
-
-@pytest.fixture
-def parameters_add_loop():
-
-    # Here we are creating dummy coordinates and flux for a loop
-    xloop = np.ones(8, dtype=np.float32) * 7
-    yloop = np.arange(11, 3, -1, dtype=np.float32)
-    zloop = np.array([1, 2, 4, 3, 4, 12, 6, 3], dtype=np.float32)
-
-    iloop = 0
-    np1 = len(xloop)
-
-    # Calculate the length of each point
-    lengths = np.zeros((np1), dtype=np.float32)
-
-    for ip in range(1, np1):
-        lengths[ip] = lengths[ip - 1] + np.sqrt(
-            (xloop[ip] - xloop[ip - 1]) ** 2 + (yloop[ip] - yloop[ip - 1]) ** 2
-        )
-
-    # The empty structures in which the first loop is stored
-    loops = []
-    loopfile = None
-
-    return (lengths, xloop, yloop, zloop, iloop, loops, loopfile)
-
-
-def test_add_loop(parameters_add_loop):
-
-    # We call the add_loop function and the values should be placed in the structures
-    loopfile, loops, iloop = loop_add(*parameters_add_loop)
-
-    expect_loopfile = np.array(
-        [
-            [0.0, 7.0, 11.0, 1.0, 0.0],
-            [0.0, 7.0, 10.0, 2.0, 1.0],
-            [0.0, 7.0, 9.0, 4.0, 2.0],
-            [0.0, 7.0, 8.0, 3.0, 3.0],
-            [0.0, 7.0, 7.0, 4.0, 4.0],
-            [0.0, 7.0, 6.0, 12.0, 5.0],
-            [0.0, 7.0, 5.0, 6.0, 6.0],
-        ]
-    )
-
-    expect_loops = [
-        [[7.0, 11.0], [7.0, 10.0], [7.0, 9.0], [7.0, 8.0], [7.0, 7.0], [7.0, 6.0], [7.0, 5.0]]
-    ]
-
-    assert np.allclose(loopfile, expect_loopfile)
-    assert np.allclose(loops, expect_loops)
-    assert np.allclose(iloop, 1)
-
-
-def test_parameters_add_loop(parameters_add_loop):
-
-    lengths, xloop, yloop, zloop, iloop, loops, loopfile = parameters_add_loop
-
-    assert np.allclose(lengths, np.arange(0, 8))
-    assert np.allclose(xloop, np.ones(8) * 7)
-    assert np.allclose(yloop, np.arange(11, 3, -1))
-    assert np.allclose(zloop, np.array([1, 2, 4, 3, 4, 12, 6, 3]))
-    assert np.allclose(iloop, 0)
-    assert not loops
-
-    assert loopfile is None
